rename llm_client module
