+++ benchmark/dbally_benchmark/e2e_benchmark.py
@@ -23,7 +23,7 @@ from sqlalchemy import create_engine
 import dbally
 from dbally.collection import Collection
 from dbally.iql_generator.iql_prompt_template import default_iql_template
-from dbally.llm_client.openai_client import OpenAIClient
+from dbally.llms.litellm import OpenAIClient
 from dbally.utils.errors import NoViewFoundError, UnsupportedQueryError
 from dbally.view_selection.view_selector_prompt_template import default_view_selector_template
