+++ benchmark/dbally_benchmark/iql_benchmark.py
@@ -22,7 +22,7 @@ from sqlalchemy import create_engine
 from dbally.audit.event_tracker import EventTracker
 from dbally.iql_generator.iql_generator import IQLGenerator
 from dbally.iql_generator.iql_prompt_template import default_iql_template
-from dbally.llm_client.openai_client import OpenAIClient
+from dbally.llms.litellm import OpenAIClient
 from dbally.utils.errors import UnsupportedQueryError
 from dbally.views.structured import BaseStructuredView
