+++ docs/quickstart/quickstart2_code.py
@@ -12,7 +12,7 @@ from dbally import decorators, SqlAlchemyBaseView
 from dbally.audit.event_handlers.cli_event_handler import CLIEventHandler
 from dbally.similarity import SimpleSqlAlchemyFetcher, FaissStore, SimilarityIndex
 from dbally.embedding_client.openai import OpenAiEmbeddingClient
-from dbally.llm_client.openai_client import OpenAIClient
+from dbally.llms.litellm import OpenAIClient
 
 engine = create_engine('sqlite:///candidates.db')
