+++ examples/recruiting.py
@@ -8,7 +8,7 @@ from recruting.views import RecruitmentView
 import dbally
 from dbally.audit.event_handlers.cli_event_handler import CLIEventHandler
 from dbally.audit.event_tracker import EventTracker
-from dbally.llm_client.openai_client import OpenAIClient
+from dbally.llms.litellm import OpenAIClient
 from dbally.prompts import PromptTemplate
 
 TEXT2SQL_PROMPT_TEMPLATE = PromptTemplate(
