+++ src/dbally/view_selection/base.py
@@ -2,7 +2,7 @@ import abc
 from typing import Dict, Optional
 
 from dbally.audit.event_tracker import EventTracker
-from dbally.llm_client.base import LLMOptions
+from dbally.llms.base import LLMOptions
 
 
 class ViewSelector(abc.ABC):
