+++ src/dbally/view_selection/random_view_selector.py
@@ -2,7 +2,7 @@ import random
 from typing import Dict, Optional
 
 from dbally.audit.event_tracker import EventTracker
-from dbally.llm_client.base import LLMOptions
+from dbally.llms.base import LLMOptions
 from dbally.view_selection.base import ViewSelector
